기초 디자인까지

+++ WebContent/jsp/usr/article/detail.jsp
@@ -1,25 +1,15 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
-<%@ page import="java.util.Map"%>
-<%@ page import="java.util.List"%>
-<%@ taglib uri="http://java.sun.com/jstl/core" prefix="c" %>
+<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
+<c:set var="pageTitle" value="게시물 상세정보" />
 
-<%@ page import="com.sbs.example.jspCommunity.dto.Article"%>
-<%
-	
-String boardName = (String) request.getAttribute("boardName");
-%>
-<!doctype html>
-<html lang="ko">
-<head>
+<%@ include file="../../part/head.jspf"%>
 
-<meta charset="UTF-8" />
-<title>게시물 정보</title>
-</head>
-<body>
+	<div class="con-min-width">
+	<div class="con">
+
 	<h1>게시물 정보</h1>
 
-	<div>
 		번호 :
 		${article.getId()}
 		<br />
@@ -30,10 +20,10 @@ String boardName = (String) request.getAttribute("boardName");
 		${article.getUpdateDate()}
 		<br />
 		게시글 분류 :
-		<%=boardName%>
+		${boardName}
 		<br />
 		작성자 :
-		${article.getExtra__writer()}
+		${article.getExtra__nickname()} 
 		<br />
 		제목 :
 		${article.getTitle()}
@@ -42,26 +32,31 @@ String boardName = (String) request.getAttribute("boardName");
 			${article.getBody()}
 		<hr />
 	</div>
-	<hr />
+	
+	
+	<div class="bot-menu con-min-width">
+	<div class="con">
 	<div>
-	<a href="detail?articleId=${article.getId() -1}">
-	&lt;이전글
+	<a href="detail?articleId=${article.getId() -1}" class="highlight">
+	&lt;이전글 |
 	</a>
-	<a href="detail?articleId=${article.getId() +1}">
+	<a href="detail?articleId=${article.getId() +1}" class="highlight">
 	다음글 &gt;
 	</a>
-	
+	<hr>
 	</div>
 
-	<div>
-		<a href="list?boardId=${article.getBoardId()}">리스트로 이동</a>
-		<a href="modify?articleId=${article.getId()}">글 수정</a>
-		<a href="delete?articleId=${article.getId()}"
-			onclick="return confirm('게시물을 삭제하시겠습니까?');">삭제</a>
+		<a href="list?boardId=${article.getBoardId()}" class="highlight">리스트 </a>|
+		
+		
+		<a href="modify?articleId=${article.getId()}" class="highlight">글 수정 </a>|
+		<a class="highlight" href="delete?articleId=${article.getId()}" onclick="return confirm('게시물을 삭제하시겠습니까?');">글 삭제</a>
 
+		
 
+</div>
 	</div>
-
+</div>
 
 
 </body>
